pyproject.toml
[tool.poetry]
name = "art-emotion-transfer"
version = "0.1.0"
description = ""
authors = ["Ruichu-Xia"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
ipykernel = "^6.29.5"
datasets = "^3.2.0"
pillow = "^11.0.0"
matplotlib = "^3.10.0"
transformers = "^4.47.1"
torch = "^2.5.1"
torchvision = "^0.20.1"
frozendict = "^2.4.6"
ruff = "^0.9.0"
pydantic-settings = "^2.7.1"

[tool.ruff]
line-length = 100
[tool.ruff.mccabe]
max-complexity = 16

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
